+++ server/services/core/src/domain/chat/chat.module.ts
@@ -1,6 +1,6 @@
 import { Logger, Module } from '@nestjs/common';
-import { ChatService } from './chat.service';
-import { ChatController } from './chat.controller';
+import { ChatService } from './service/chat.service';
+import { ChatController } from './controller/chat.controller';
 
 @Module({
   controllers: [ChatController],
