+++ server/services/core/src/domain/fanup/controller/fanup.controller.ts
@@ -1,8 +1,8 @@
 import { Controller } from '@nestjs/common';
 import { MessagePattern, Payload } from '@nestjs/microservices';
-import { FanupService } from './fanup.service';
-import { CreateFanupDto } from './dto/create-fanup.dto';
-import { UpdateFanupDto } from './dto/update-fanup.dto';
+import { FanupService } from '../service/fanup.service';
+import { CreateFanupDto } from '../dto/create-fanup.dto';
+import { UpdateFanupDto } from '../dto/update-fanup.dto';
 
 @Controller()
 export class FanupController {
