+++ server/services/core/src/domain/fanup/service/fanup.service.spec.ts
